requirements.txt
auto-subtitle
ffmpeg-python
pyannote-audio